Remove user account.

+++ RewriteMe.Mobile/Apps/RewriteMe.Mobile/ViewModels/UserSettingsPageViewModel.cs
@@ -1,4 +1,5 @@
-﻿using System.Threading.Tasks;
+﻿using System;
+using System.Threading.Tasks;
 using System.Windows.Input;
 using Prism.Navigation;
 using RewriteMe.Domain.Http;
@@ -49,9 +50,37 @@ namespace RewriteMe.Mobile.ViewModels
 
         private async Task ExecuteDeleteAccountCommandAsync()
         {
+            var userSession = await UserSessionService.GetUserSessionAsync().ConfigureAwait(false);
+            var message = Loc.Text(TranslationKeys.DeleteUserAccountMessage, userSession.Email);
+            var result = await DialogService.PromptAsync(
+                message,
+                okText: Loc.Text(TranslationKeys.Ok),
+                cancelText: Loc.Text(TranslationKeys.Cancel)).ConfigureAwait(false);
+
+            if (!result.Ok)
+                return;
+
+            if (!result.Text.Equals(userSession.Email, StringComparison.OrdinalIgnoreCase))
+            {
+                await DialogService.AlertAsync(
+                    Loc.Text(TranslationKeys.IncorrectUserInputErrorMessage),
+                    Loc.Text(TranslationKeys.Warning),
+                    Loc.Text(TranslationKeys.Ok)).ConfigureAwait(false);
+                return;
+            }
+
             var httpRequestResult = await _rewriteMeWebService.DeleteUserAsync().ConfigureAwait(false);
-            if (httpRequestResult.State == HttpRequestState.Success)
-            { }
+            if (httpRequestResult.State != HttpRequestState.Success)
+            {
+                await DialogService.AlertAsync(
+                    Loc.Text(TranslationKeys.UserAccountDeletionErrorMessage),
+                    Loc.Text(TranslationKeys.Warning),
+                    Loc.Text(TranslationKeys.Ok)).ConfigureAwait(false);
+                return;
+            }
+
+            await UserSessionService.SignOutAsync().ConfigureAwait(false);
+            await NavigationService.NavigateWithoutAnimationAsync($"/{Pages.Navigation}/{Pages.Login}").ConfigureAwait(false);
         }
 
         private async Task ExecuteLogoutCommandAsync()
